// src/app/api/models/applicant-summary-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface ApplicantSummaryDto {
  aiMatchScore?: number;
  applicationDate?: string;
  candidateName?: string;
  experienceYears?: number;
  id?: string;
  jobId?: string;
  jobTitle?: string;
  location?: string;
  profileImageUrl?: string;
  skills?: Array<string>;
  status?: string;
}